pyproject.toml
[project]
name = "tg-parser"
version = "0.1.0"
description = "App for propaganda-spreading bots detection"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
packages = [
    { include = "tg_parser", from = "." }
]


requires-python = ">=3.10"
dependencies = [
    "telethon (>=1.38.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.3,<3.0.0)",
    "emoji (>=2.14.1,<3.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "datasets (>=3.3.2,<4.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "streamlit (>=1.42.2,<2.0.0)",
    "xgboost (>=2.1.4,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "flake8 (>=7.1.2,<8.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
